+++ CHANGELOG.rst
@@ -38,6 +38,8 @@ Thanks :user:`ddelange` for the PR.
 Previously-deprecated API have been removed, including:
 
 - `default` and `missing` parameters, which were replaced by `dump_default` and `load_default` in 3.13.0 (:pr:`1742`, :pr:`2700`).
+- Passing field metadata via keyword arguments is no longer supported (deprecated in 3.10.0). Use the explicit ``metadata=...``
+  argument instead (:issue:`1350`).
 
 3.23.2 (2024-12-18)
 *******************